pyproject.toml
[project]
name = "10minagent"
version = "0.1.0"
description = "Dr. KAI - AI Medical Assistant with approved Swedish sources"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "google-trans-new>=1.1.9",
    "langdetect>=1.0.9",
    "serpapi>=0.1.5",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
]

[project.scripts]
drkai = "main:main"
drkai-api = "api:app"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
